+++ src/clients/python/experimental_api_v2/library/cuda_shared_memory/cuda_shared_memory.cc
@@ -110,11 +110,11 @@ CudaSharedMemoryGetRawHandle(
   base64_init_encodestate(&es);
   size_t handle_size = sizeof(cudaIpcMemHandle_t);
   *serialized_raw_handle = (char*)malloc(handle_size * 2); /* ~4/3 x input */
-  size_t offset = base64_encode_block(
+  int offset = base64_encode_block(
       (char*)((void*)&handle->cuda_shm_handle_), handle_size,
       *serialized_raw_handle, &es);
   base64_encode_blockend(*serialized_raw_handle + offset, &es);
-  size_t padding_size =
+  int padding_size =
       base64_encode_blockend(*serialized_raw_handle + offset, &es);
   offset += (padding_size - 1);
   // The base64_encode_blockend does not null-terminate the string but adds
